atulização da tela busca de cliente

+++ Interações/Servidor/cadastrar.py
@@ -876,18 +876,22 @@ ENGINE = InnoDB;"""
                 #print(lista)
                 print('cpf:',cpf)
                 cliente = self.sqlite_readSec_cliente(cpf)
-                print('cliente[0][0]:',cliente[0][0])
-                cursor.execute('SELECT * FROM `Vendas` WHERE Cliente_idCliente= %s'%cliente[0][0])
-                venda = cursor.fetchall()
-                print('Venda:',venda)
-                if venda:
-                    print("Falha ao remover cliente: Registro de venda feita por esse cliente.")
-                    return False
+                if cliente:
+                    print('cliente[0][0]:',cliente[0][0])
+                    cursor.execute('SELECT * FROM `Vendas` WHERE Cliente_idCliente= %s'%cliente[0][0])
+                    venda = cursor.fetchall()
+                    print('Venda:',venda)
+                    if venda:
+                        print("Falha ao remover cliente: Registro de venda feita por esse cliente.")
+                        return False
+                    else:
+                        cursor.execute('DELETE FROM `mydb_2`.`cliente` WHERE CPF = %s' % (str(cpf)))
+                        conexao.commit()
+                        conexao.close()
+                        return True
                 else:
-                    cursor.execute('DELETE FROM `mydb_2`.`cliente` WHERE CPF = %s' % (str(cpf)))
-                    conexao.commit()
-                    conexao.close()
-                    return True
+                    print('Falha ao remover cliente: Registro de cliente não encontrado.')
+                    return False
 
 
         except Error as erro:
